small changes in admin area

+++ project/website/admin.py
@@ -1,31 +1,30 @@
 from flask import Flask, Blueprint, render_template, request, flash, redirect, url_for
 from flask_login import login_required, current_user
 from .models import User, Seat
+from .python_scripts import admin_user_check
 from . import db
 from .CharReader import Dictionary_Creater
 from .CharReader import model_seat_filler
 import os
+
 
 admins = Blueprint('admins', __name__)
 
 @admins.route('/', methods=['GET','POST'])
 @login_required
 def admin_home():
-    if current_user.role == 'admin':
+    Admin = admin_user_check(current_user)
+    if Admin:
         return render_template('admin_home.html', user= current_user)
-    else:
-        flash('You need do be an Admin to access!!', category='error')
-        return redirect(url_for('views.home'))
+
 
 @admins.route('/user', methods=['GET', 'POST'])
 @login_required
 def admin_user():
-    if current_user.role == 'admin':
+    Admin = admin_user_check(current_user)
+    if Admin:
         users = User.query.all()
         return render_template('admin_user.html', user=current_user, users=users)
-    else:
-        flash('You need do be an Admin to access!!', category='error')
-        return redirect(url_for('views.home'))
 
 @admins.route('/user/update-user/<int:id>', methods=['GET', 'POST'])
 @login_required
@@ -49,42 +48,42 @@ def update_user(id):
 @admins.route('user/delete-user/<int:id>',methods=['POST', 'GET'])
 @login_required
 def delete_user(id):
-    user_to_delete = User.query.get_or_404(id)
-    try:
-        db.session.delete(user_to_delete)
-        db.session.commit()
-        flash('Account deleted!', category='success')
-        return redirect('/admin-area/user')
-    except:
-        flash('Account NOT deleted!', category='error')
-        return redirect('/admin-area/user')
+    Admin = admin_user_check(current_user)
+    if Admin:
+        user_to_delete = User.query.get_or_404(id)
+        try:
+            db.session.delete(user_to_delete)
+            db.session.commit()
+            flash('Account deleted!', category='success')
+            return redirect('/admin-area/user')
+        except:
+            flash('Account NOT deleted!', category='error')
+            return redirect('/admin-area/user')
 
 @admins.route('/update-flights', methods=['GET','POST'])
 @login_required
 def admin_flights():
-    if current_user.role == 'admin':
+    Admin = admin_user_check(current_user)
+    if Admin:
         Path = os.path.abspath(os.curdir)
         ChartIn_Path = Path +'\Input_Data\\'
         Flight_Dictionary = Dictionary_Creater(ChartIn_Path)
         model_seat_filler(Flight_Dictionary)
         return redirect('/admin-area/seats')
-    else:
-        flash('You need do be an Admin to access!!', category='error')
-        return redirect(url_for('views.home'))
 
 @admins.route('/seats', methods=['GET', 'POST'])
 @login_required
 def admin_seats():
-    if current_user.role == 'admin':
+    Admin = admin_user_check(current_user)
+    if Admin:
         seats = Seat.query.all()
         return render_template('admin_seats.html', user=current_user, seats=seats)
-    else:
-        flash('You need do be an Admin to access!!', category='error')
-        return redirect(url_for('views.home'))
+
 @admins.route('/seats/update-seats/<int:id>', methods=['GET','POST'])
 @login_required
 def update_seat(id):
-    if current_user.role == 'admin':
+    Admin = admin_user_check(current_user)
+    if Admin:
         seat_to_update = Seat.query.get_or_404(id)
         if request.method == 'POST':
             seat_to_update.user_id = request.form['user_id']
@@ -101,8 +100,6 @@ def update_seat(id):
                 return redirect('/admin-area/seats')
         else:
             return render_template('admin_update_seat.html', user= current_user, seat_to_update=seat_to_update)
-    else:
-        flash('You need do be an Admin to access!!', category='error')
-        return redirect(url_for('views.home'))
+
